refactor: 파일 단순화
