--- src/basic/components/index.js
@@ -1,5 +1,0 @@
-/**
- * 컴포넌트 Barrel Exports
- * React 스타일의 컴포넌트 시스템 준비
- */
-export { Button, AddToCartButton, QuantityButton } from './ui/button.js';
