--- src/basic/constants/index.js
@@ -1,8 +1,0 @@
-// 🔥 Barrel exports pattern - 모든 상수를 한 곳에서 import 가능
-export * from './products.js';
-export * from './discounts.js';
-export * from './points.js';
-
-// 기타 상수
-export const LOW_STOCK_THRESHOLD = 5;
-export const TOTAL_STOCK_WARNING_THRESHOLD = 50;
