--- src/basic/constants/points.js
@@ -1,12 +1,0 @@
-// 포인트 시스템 상수
-export const POINTS_CALCULATION_BASE = 1000;
-export const COMBO_BONUS_POINTS = 50;
-export const FULL_SET_BONUS_POINTS = 100;
-export const SMALL_BULK_BONUS_POINTS = 20;
-export const MEDIUM_BULK_BONUS_POINTS = 50;
-export const LARGE_BULK_BONUS_POINTS = 100;
-
-// 구매 수량 임계값
-export const SMALL_BULK_THRESHOLD = 10;
-export const MEDIUM_BULK_THRESHOLD = 20;
-export const LARGE_BULK_THRESHOLD = 30;
