--- src/basic/features/cart/templates/index.js
@@ -1,4 +1,0 @@
-/**
- * 장바구니 템플릿 Barrel Exports
- */
-export { cartItemTemplate } from './cart-item.js';
