--- src/basic/features/discounts/index.js
@@ -1,5 +1,0 @@
-// Discounts Feature 진입점  
-export { getProductDiscountRate, applyProductDiscount } from './strategies/product-discount.js';
-export { applyBulkDiscount } from './strategies/bulk-discount.js';
-export { calculateTuesdayDiscount } from './strategies/tuesday-discount.js';
-export { updateTuesdayUI } from './ui/discount-ui.js';
