--- src/basic/features/events/index.js
@@ -1,2 +1,0 @@
-// Events Feature 진입점
-export { setupEventTimers } from './timers/event-timers.js';
