--- src/basic/features/points/index.js
@@ -1,13 +1,0 @@
-// Points Feature 진입점
-export { 
-  calculateBasePoints, 
-  calculateTuesdayBonus, 
-  calculateComboBonuses, 
-  calculateBulkBonus 
-} from './calculators/points-calculator.js';
-
-export { 
-  displayPointsInfo, 
-  hidePointsIfEmpty,
-  renderBonusPoints
-} from './ui/points-display.js';
