--- src/basic/features/products/index.js
@@ -1,3 +1,0 @@
-// Products Feature 진입점
-export { initializeProducts, findProduct, getLowStockProducts, findProductById } from './data/product-manager.js';
-export { getOptionData, getProductDisplayInfo, updateSelectOptions } from './ui/product-options.js';
