--- src/basic/features/products/templates/index.js
@@ -1,1 +1,0 @@
-export { productSelectorTemplate } from './product-selector.js';
