+++ src/basic/main-old.js
@@ -1,137 +1,71 @@
-/**
- * Clean Shopping Cart Application - 단순화된 구조
- * 모든 shared 유틸리티와 constants가 직접 포함됨
- */
 
-// ==================== Constants (분산 제거) ====================
 
-// Products
-const PRODUCT_ONE = 'p1';
-const PRODUCT_TWO = 'p2'; 
-const PRODUCT_THREE = 'p3';
-const PRODUCT_FOUR = 'p4';
-const PRODUCT_FIVE = 'p5';
+const {
+  QUANTITY_DISCOUNT_THRESHOLD
+} = constants;
 
-const INITIAL_PRODUCT_DATA = [
-  { id: PRODUCT_ONE, name: '버그 없애는 키보드', price: 10000, originalPrice: 10000, quantity: 50, onSale: false, suggestSale: false },
-  { id: PRODUCT_TWO, name: '생산성 폭발 마우스', price: 20000, originalPrice: 20000, quantity: 30, onSale: false, suggestSale: false },
-  { id: PRODUCT_THREE, name: '거북목 탈출 모니터암', price: 30000, originalPrice: 30000, quantity: 20, onSale: false, suggestSale: false },
-  { id: PRODUCT_FOUR, name: '에러 방지 노트북 파우치', price: 15000, originalPrice: 15000, quantity: 0, onSale: false, suggestSale: false },
-  { id: PRODUCT_FIVE, name: '코딩할 때 듣는 Lo-Fi 스피커', price: 25000, originalPrice: 25000, quantity: 10, onSale: false, suggestSale: false }
-];
-
-// Discounts
-const KEYBOARD_DISCOUNT_RATE = 0.1;
-const MOUSE_DISCOUNT_RATE = 0.15;
-const MONITOR_ARM_DISCOUNT_RATE = 0.2;
-const SPEAKER_DISCOUNT_RATE = 0.25;
-const BULK_DISCOUNT_THRESHOLD = 30;
-const BULK_DISCOUNT_RATE = 0.25;
-const QUANTITY_DISCOUNT_THRESHOLD = 10;
-
-// Times & Events
-const TUESDAY_DAY_NUMBER = 2;
-const TUESDAY_ADDITIONAL_DISCOUNT_RATE = 0.1;
-const LIGHTNING_SALE_MAX_DELAY = 10000;
-const LIGHTNING_SALE_DISCOUNT_RATE = 0.2;
-const LIGHTNING_SALE_DURATION = 30000;
-const SUGGESTION_DISCOUNT_RATE = 0.05;
-const SUGGESTION_SALE_MAX_DELAY = 20000;
-const SUGGESTION_INTERVAL_MS = 60000;
-
-// Points
-const POINTS_CALCULATION_BASE = 1000;
-const COMBO_BONUS_POINTS = 50;
-const FULL_SET_BONUS_POINTS = 100;
-const SMALL_BULK_BONUS_POINTS = 20;
-const MEDIUM_BULK_BONUS_POINTS = 50;
-const LARGE_BULK_BONUS_POINTS = 100;
-const SMALL_BULK_THRESHOLD = 10;
-const MEDIUM_BULK_THRESHOLD = 20;
-const LARGE_BULK_THRESHOLD = 30;
-
-// General
-const LOW_STOCK_THRESHOLD = 5;
-const TOTAL_STOCK_WARNING_THRESHOLD = 50;
-
-// Sale texts
-const SUPER_SALE_TEXT = '25% SUPER SALE!';
-const LIGHTNING_SALE_TEXT = '20% SALE!';
-const SUGGESTION_SALE_TEXT = '5% 추천할인!';
-
-// ==================== Imports ====================
-
-import { initializeProducts, findProductById, getProductDisplayInfo, updateSelectOptions, productSelectorTemplate } from './features/products.js';
-import { updateTuesdayUI } from './features/discounts.js';
-import { getCartProductTypes, updatePricesInCart, calculateCompleteCartTotals, updateCartTotalsDisplay, handleAddToCart, handleCartActions, cartItemTemplate } from './features/cart.js';
-import { calculateBasePoints, calculateTuesdayBonus, calculateComboBonuses, calculateBulkBonus, renderBonusPoints } from './features/points.js';
-import { setupEventTimers } from './features/events.js';
-
-// ==================== App State ====================
-
-const appState = {
-  totalAmount: 0,
-  itemCount: 0,
-  lastSelectedItem: null,
-  bonusPoints: 0
-};
-
-const dataState = {
-  productList: null
-};
-
-const domRefs = {
-  stockInformation: null,
-  productSelect: null,
-  addButton: null,
-  cartDisplay: null,
-  cartTotalElement: null
-};
-
-// ==================== Shared Utilities (직접 포함) ====================
+// ==================== Shared Utilities (통합) ====================
 
 const specialProperties = ["checked", "selected", "disabled", "readOnly"];
 
 /**
  * DOM 요소 생성 헬퍼
+ * @param {string} tagName - 태그명
+ * @param {Object} props - 속성 객체
+ * @param {...any} children - 자식 요소들
+ * @returns {Element} DOM 요소
  */
 const createElement = (tagName, props = {}, ...children) => {
   const element = document.createElement(tagName);
 
   if (props) {
-    Object.entries(props).forEach(([key, value]) => {
-      if (key.startsWith("on") && typeof value === "function") {
-        const eventType = key.slice(2).toLowerCase();
-        element.addEventListener(eventType, value);
-      } else if (key === "className") {
-        element.className = value;
-      } else if (key === "innerHTML") {
-        element.innerHTML = value;
-      } else if (key === "textContent") {
-        element.textContent = value;
-      } else if (specialProperties.includes(key)) {
-        element[key] = value;
-      } else {
-        element.setAttribute(key, value);
-      }
-    });
+    applyProps(element, props);
   }
 
   if (children && children.length > 0) {
-    children.forEach(child => {
-      if (child === null || child === undefined) return;
-      
-      if (typeof child === "string" || typeof child === "number") {
-        element.appendChild(document.createTextNode(String(child)));
-      } else if (Array.isArray(child)) {
-        child.forEach(c => element.appendChild(c));
-      } else if (child && child.nodeType === 1) {
-        element.appendChild(child);
-      }
-    });
+    appendChildren(element, children);
   }
 
   return element;
+};
+
+/**
+ * 요소에 props 적용
+ */
+const applyProps = (element, props) => {
+  Object.entries(props).forEach(([key, value]) => {
+    if (key.startsWith("on") && typeof value === "function") {
+      const eventType = key.slice(2).toLowerCase();
+      element.addEventListener(eventType, value);
+    } else if (key === "className") {
+      element.className = value;
+    } else if (key === "innerHTML") {
+      element.innerHTML = value;
+    } else if (key === "textContent") {
+      element.textContent = value;
+    } else if (specialProperties.includes(key)) {
+      element[key] = value;
+    } else {
+      element.setAttribute(key, value);
+    }
+  });
+};
+
+/**
+ * 자식 요소들 추가
+ */
+const appendChildren = (element, children) => {
+  children.forEach(child => {
+    if (child === null || child === undefined) return;
+    
+    if (typeof child === "string" || typeof child === "number") {
+      element.appendChild(document.createTextNode(String(child)));
+    } else if (Array.isArray(child)) {
+      appendChildren(element, child);
+    } else if (child && child.nodeType === 1) {
+      element.appendChild(child);
+    }
+  });
 };
 
 /**
@@ -149,6 +83,8 @@ const setDOMRefs = (target, selectors) => {
 
 /**
  * 장바구니 아이템의 스타일을 업데이트합니다
+ * @param {HTMLElement} cartItem - 장바구니 DOM 요소
+ * @param {number} quantity - 상품 수량
  */
 const updateItemStyles = (cartItem, quantity) => {
   const priceElements = cartItem.querySelectorAll(".text-lg, .text-xs");
@@ -159,8 +95,46 @@ const updateItemStyles = (cartItem, quantity) => {
   });
 };
 
-// ==================== Templates (직접 포함) ====================
+/**
+ * 매뉴얼 시스템 컴포넌트
+ * @param {Object} dependencies - 필요한 의존성들
+ * @returns {Object} 매뉴얼 시스템 요소들
+ */
+const createManualSystem = (dependencies) => {
+  const { createElement, manualGuideTemplate, helpToggleTemplate } = dependencies;
 
+  const manualColumn = createElement("div", {
+    className: "fixed right-0 top-0 h-full w-80 bg-white shadow-2xl p-6 overflow-y-auto z-50 transform translate-x-full transition-transform duration-300",
+    innerHTML: manualGuideTemplate
+  });
+
+  const manualOverlay = createElement("div", {
+    className: "fixed inset-0 bg-black/50 z-40 hidden transition-opacity duration-300",
+    onclick: (event) => {
+      if (event.target === manualOverlay) {
+        manualOverlay.classList.add("hidden");
+        manualColumn.classList.add("translate-x-full");
+      }
+    }
+  }, [manualColumn]);
+
+  const manualToggle = createElement("button", {
+    className: "fixed top-4 right-4 bg-black text-white p-3 rounded-full hover:bg-gray-900 transition-colors z-50",
+    innerHTML: helpToggleTemplate,
+    onclick: () => {
+      manualOverlay.classList.toggle("hidden");
+      manualColumn.classList.toggle("translate-x-full");
+    }
+  });
+
+  return { manualToggle, manualOverlay };
+};
+
+// ==================== Templates (통합) ====================
+
+/**
+ * 헤더 템플릿
+ */
 const headerTemplate = `
   <div class="mb-8">
     <h1 class="text-xs font-medium tracking-extra-wide uppercase mb-2">🛒 Hanghae Online Store</h1>
@@ -169,6 +143,9 @@ const headerTemplate = `
   </div>
 `;
 
+/**
+ * 주문 요약 템플릿
+ */
 const orderSummaryTemplate = `
   <div class="bg-black text-white p-8 flex flex-col">
     <h2 class="text-xs font-medium mb-5 tracking-extra-wide uppercase">Order Summary</h2>
@@ -199,15 +176,21 @@ const orderSummaryTemplate = `
       <span id="points-notice">Earn loyalty points with purchase.</span>
     </p>
   </div>
-`;
+\`;
 
-const helpToggleTemplate = `
+/**
+ * 도움말 토글 버튼 템플릿
+ */
+const helpToggleTemplate = \`
 <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
   <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path>
 </svg>
-`;
+\`;
 
-const manualGuideTemplate = `
+/**
+ * 이용 안내 매뉴얼 템플릿
+ */
+const manualGuideTemplate = \`
 <button class="absolute top-4 right-4 text-gray-500 hover:text-black" onclick="document.querySelector('.fixed.inset-0').classList.add('hidden'); this.parentElement.classList.add('translate-x-full')">
   <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
     <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"></path>
@@ -272,59 +255,51 @@ const manualGuideTemplate = `
     • 상품4 = 품절
   </p>
 </div>
-`;
+\`;
 
-const cartSummaryTemplate = (summaryItems, subtotal) => `
-${summaryItems}
+/**
+ * 장바구니 요약 정보 템플릿 (동적)
+ */
+const cartSummaryTemplate = (summaryItems, subtotal) => \`
+\${summaryItems}
 <div class="border-t border-gray-200 my-2"></div>
 <div class="flex justify-between text-sm font-medium">
   <span>소계</span>
-  <span>₩${subtotal.toLocaleString()}</span>
+  <span>₩\${subtotal.toLocaleString()}</span>
 </div>
-`;
+\`;
 
-const discountInfoTemplate = (discRate, savedAmount) => `
+/**
+ * 할인 정보 템플릿 (동적)
+ */
+const discountInfoTemplate = (discRate, savedAmount) => \`
 <div class="bg-green-500/20 rounded-lg p-3">
   <div class="flex justify-between items-center mb-1">
     <span class="text-xs uppercase tracking-wide text-green-400">총 할인율</span>
-    <span class="text-sm font-medium text-green-400">${(discRate * 100).toFixed(1)}%</span>
+    <span class="text-sm font-medium text-green-400">\${(discRate * 100).toFixed(1)}%</span>
   </div>
-  <div class="text-2xs text-gray-300">₩${Math.round(savedAmount).toLocaleString()} 할인되었습니다</div>
+  <div class="text-2xs text-gray-300">₩\${Math.round(savedAmount).toLocaleString()} 할인되었습니다</div>
 </div>
-`;
+\`;
 
-/**
- * 매뉴얼 시스템 생성
- */
-const createManualSystem = () => {
-  const manualColumn = createElement("div", {
-    className: "fixed right-0 top-0 h-full w-80 bg-white shadow-2xl p-6 overflow-y-auto z-50 transform translate-x-full transition-transform duration-300",
-    innerHTML: manualGuideTemplate
-  });
-
-  const manualOverlay = createElement("div", {
-    className: "fixed inset-0 bg-black/50 z-40 hidden transition-opacity duration-300",
-    onclick: (event) => {
-      if (event.target === manualOverlay) {
-        manualOverlay.classList.add("hidden");
-        manualColumn.classList.add("translate-x-full");
-      }
-    }
-  }, [manualColumn]);
-
-  const manualToggle = createElement("button", {
-    className: "fixed top-4 right-4 bg-black text-white p-3 rounded-full hover:bg-gray-900 transition-colors z-50",
-    innerHTML: helpToggleTemplate,
-    onclick: () => {
-      manualOverlay.classList.toggle("hidden");
-      manualColumn.classList.toggle("translate-x-full");
-    }
-  });
-
-  return { manualToggle, manualOverlay };
+const appState = {
+  totalAmount: 0,
+  itemCount: 0,
+  lastSelectedItem: null,
+  bonusPoints: 0
 };
 
-// ==================== App Logic ====================
+const dataState = {
+  productList: null
+};
+
+const domRefs = {
+  stockInformation: null,
+  productSelect: null,
+  addButton: null,
+  cartDisplay: null,
+  cartTotalElement: null
+};
 
 const initializeAppState = () => {
   appState.totalAmount = 0;
@@ -333,7 +308,7 @@ const initializeAppState = () => {
 };
 
 const initializeProductData = () => {
-  dataState.productList = [...INITIAL_PRODUCT_DATA];
+  dataState.productList = initializeProducts();
 };
 
 const main = () => {
@@ -384,7 +359,11 @@ const createDOMStructure = (root) => {
     className: "grid grid-cols-1 lg:grid-cols-[1fr_360px] gap-6 flex-1 overflow-hidden" 
   }, [leftColumn, rightColumn]);
 
-  const { manualToggle, manualOverlay } = createManualSystem();
+  const { manualToggle, manualOverlay } = createManualSystem({
+    createElement,
+    manualGuideTemplate,
+    helpToggleTemplate
+  });
 
   const appStructure = createElement("div", { className: "app-container" }, [
     header,
@@ -399,11 +378,19 @@ const createDOMStructure = (root) => {
 };
 
 /**
- * 장바구니 총액 계산 및 UI 업데이트
+ * 장바구니 총액 계산 및 UI 업데이트 - 조합 함수 (분리 완료)
+ * 
+ * 추상화 수준 일관성 개선:
+ * - 비즈니스 로직: cart/business/calculations.js → calculateCompleteCartTotals()  
+ * - UI 업데이트: cart/ui/dom-updates.js → updateCartTotalsDisplay()
+ * - 조합 로직: main.basic.js → 단순한 함수 호출 조합
+ * 
+ * 이제 다른 복잡한 함수들과 동일한 추상화 수준 달성! ✅
  */
 const calculateCartTotals = () => {
   const cartItems = domRefs.cartDisplay.children;
   
+  // 1. 순수한 계산 수행
   const calculationResult = calculateCompleteCartTotals(
     cartItems, 
     dataState.productList, 
@@ -415,6 +402,7 @@ const calculateCartTotals = () => {
     }
   );
 
+  // 2. UI 업데이트 수행
   updateCartTotalsDisplay(
     calculationResult,
     domRefs,
@@ -437,6 +425,14 @@ const calculateCartTotals = () => {
 
 main();
 
+/**
+ * 이벤트 핸들러 설정 - 조합 함수로 추상화 수준 일관성 달성 ✅
+ * 
+ * 추상화 개선:
+ * - 비즈니스 로직: cart/handlers/cart-handlers.js → 순수함수로 분리
+ * - 이벤트 등록: main.basic.js → 단순한 함수 호출 조합
+ */
+
 // 장바구니 아이템 추가 이벤트
 domRefs.addButton.addEventListener("click", () => {
   const selectedItemId = domRefs.productSelect.value;
@@ -450,7 +446,7 @@ domRefs.addButton.addEventListener("click", () => {
   });
 });
 
-// 장바구니 액션 처리 이벤트
+// 장바구니 액션 처리 이벤트 (수량 변경/제거)
 domRefs.cartDisplay.addEventListener("click", (event) => {
   handleCartActions(event, {
     productList: dataState.productList,
@@ -458,4 +454,5 @@ domRefs.cartDisplay.addEventListener("click", (event) => {
     constants: { TOTAL_STOCK_WARNING_THRESHOLD },
     functions: { findProductById, calculateCartTotals, updateSelectOptions, createElement }
   });
-}); 
+});
+`;
