--- src/basic/shared/index.js
@@ -1,4 +1,0 @@
-// Shared 모듈 진입점 (도메인에 속하지 않는 공통 유틸리티)
-export { updateItemStyles } from './ui/item-styles.js';
-export { createElement, setDOMRefs, setDOMRefsFromContainer, addEventListeners } from './utils/unified-dom.js';
-export { createManualSystem } from './components/manual-system.js';
