--- src/basic/shared/templates/header.js
@@ -1,7 +1,0 @@
-export const headerTemplate = `
-  <div class="mb-8">
-    <h1 class="text-xs font-medium tracking-extra-wide uppercase mb-2">🛒 Hanghae Online Store</h1>
-    <div class="text-5xl tracking-tight leading-none">Shopping Cart</div>
-    <p id="item-count" class="text-sm text-gray-500 font-normal mt-3">🛍️ 0 items in cart</p>
-  </div>
-`;
