--- src/basic/shared/templates/help-toggle.js
@@ -1,8 +1,0 @@
-/**
- * 도움말 토글 버튼 템플릿
- */
-export const helpToggleTemplate = `
-<svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path>
-</svg>
-`;
