--- src/basic/shared/templates/index.js
@@ -1,5 +1,0 @@
-export { headerTemplate } from './header.js';
-export { orderSummaryTemplate } from './order-summary.js';
-export { manualGuideTemplate } from './manual-guide.js';
-export { helpToggleTemplate } from './help-toggle.js';
-export { cartSummaryTemplate, discountInfoTemplate } from './cart-summary.js';
